Moved some tests and moved math.js 37 / 66 tests work.
